Optional client UI rework
Remove `FORMIO_CLIENT_UI` and use `NO_INSTALL` instead.
The reason to remove is to make use of the existing
env for this functionality.
Modified server.js to address the edge case caused by enabling
the variable. Since the variable is not exposed, the maintainers
might not came accross this case or they might never intend to
even use the same, I am not sure here..., but the changes will
cover our needs without having to introduce another env.